fix for `localizedName`
